requirements.txt
streamlit>=1.32.0
pandas>=2.2.1
numpy<2.0.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
plotly>=5.18.0
supabase>=2.3.0
postgrest>=0.11.0
sqlite-utils>=3.36.0
langgraph>=0.2.0
